Refactor to use new formatters
